requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cloudinary==1.26.0
coreapi==2.3.3
cores==0.0.0.0a0.dev20210616
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
jsonschema==4.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.2
Pillow==8.4.0
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
sentry-sdk==1.4.3
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
swagger-spec-validator==2.7.3
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
